Fix issues reported by Travis.

+++ src/server/scripts/Kalimdor/CavernsOfTime/CullingOfStratholme/boss_mal_ganis.cpp
@@ -168,7 +168,7 @@ class boss_mal_ganis : public CreatureScript
                             events.Repeat(Seconds(6));
                             break;
                         case EVENT_MIND_BLAST:
-                            if (Unit* target = SelectTarget(SELECT_TARGET_RANDOM, 1, 100.0f, true, -sSpellMgr->GetSpellIdForDifficulty(SPELL_SLEEP, me)))
+                            if (Unit* target = SelectTarget(SELECT_TARGET_RANDOM, 1, 100.0f, true, -int32(sSpellMgr->GetSpellIdForDifficulty(SPELL_SLEEP, me))))
                                 DoCast(target, SPELL_MIND_BLAST);
                             else
                                 DoCastVictim(SPELL_MIND_BLAST);
